backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
Pillow==10.1.0
aiofiles==23.2.1
google-generativeai==0.3.2
google-auth==2.23.4
